Edits to index, team, carousel, and footer
Changed the lab description on the home page to black text on white background. Made the top margin of the footer smaller on all pages. Replaced bullet points on team page for Google Scholar and CV with icons. Image carousel now has a fixed aspect ratio and image width fills up the whole page while the height is cropped.